refactoring to re-use existing components, deleting unrequired lines & using formstate for errors (not yet working)
