+++ admin/resources/js/components/form/Select.tsx
@@ -1,7 +1,10 @@
-import * as React from "react";
+import React from "react";
 import { RegisterOptions, useFormContext } from "react-hook-form";
+import get from "lodash/get";
+import { InputWrapper } from "../H2Components/InputWrapper";
 
-interface SelectProps extends React.SelectHTMLAttributes<HTMLSelectElement> {
+export interface SelectProps
+  extends React.SelectHTMLAttributes<HTMLSelectElement> {
   /** HTML id used to identify the element. */
   id: string;
   /** The text for the label associated with the select input. */
@@ -12,6 +15,8 @@ interface SelectProps extends React.SelectHTMLAttributes<HTMLSelectElement> {
   options: { value: string | number; text: string }[];
   /** Object set of validation rules to impose on input. */
   rules?: RegisterOptions;
+  /** Optional context which user can view by toggling a button. */
+  context?: string;
 }
 
 const Select: React.FunctionComponent<SelectProps> = ({
@@ -20,29 +25,43 @@ const Select: React.FunctionComponent<SelectProps> = ({
   name,
   options,
   rules,
+  context,
   ...rest
 }) => {
   const {
     register,
     formState: { errors },
   } = useFormContext();
-  const error = errors[name]?.message;
+
+  const error = get(errors, name)?.message;
   return (
     <div>
-      <label htmlFor={id}>{label}</label>
-      <select
-        id={id}
-        {...register(name, rules)}
-        aria-invalid={error ? "true" : "false"}
-        {...rest}
+      <InputWrapper
+        inputId={id}
+        label={label}
+        required={!!rules?.required}
+        context={context}
+        hideOptional
       >
-        {options.map(({ value, text }) => (
-          <option key={value} value={value}>
-            {text}
-          </option>
-        ))}
-      </select>
-      {error && <span role="alert">{error}</span>}
+        <select
+          data-h2-radius="b(s)"
+          data-h2-padding="b(top-bottom, xxs) b(left, xxs)"
+          data-h2-margin="b(bottom, xs)"
+          data-h2-font-size="b(normal)"
+          id={id}
+          style={{ width: "100%" }}
+          {...register(name, rules)}
+          aria-invalid={error ? "true" : "false"}
+          {...rest}
+        >
+          <option value="" aria-label="Empty option" />
+          {options.map(({ value, text }) => (
+            <option data-h2-font-size="b(caption)" key={value} value={value}>
+              {text}
+            </option>
+          ))}
+        </select>
+      </InputWrapper>
     </div>
   );
 };
